requirements.txt
# NOTE: discord.py[voice]==2.3.2 voice client is broken on Python 3.13+.
# Install from git main for local dev on Python 3.13/3.14:
#   pip install -U "discord.py[voice] @ git+https://github.com/Rapptz/discord.py.git"
# The Dockerfile uses Python 3.11-slim where 2.3.2 works fine for server deploy.
discord.py[voice] @ git+https://github.com/Rapptz/discord.py.git
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
PyNaCl>=1.5.0
aiohttp>=3.9.0
audioop-lts>=0.2.1  # Required for Python 3.13+ (audioop was removed from stdlib)
